Changes to UI and user context
Changes to UI to replace my name with my student number, and better text presented to participants.

Changes to the user context variables and naming throughout the codebase to introduce participant_code instead of user_name. Altered query_rewriter prompt to ignore the participant_code as this is only needed for logging purposes.